modify app props, rename server renderer file
